Move SeriesModel and ParallelModel into experimental package
The series/parallel combinator classes were unused by the rest of the
package (ParallelModel was not even exported from the parametric
namespace). Relocate them under surpyval.experimental alongside the other
experimental features, export them from experimental/__init__, and switch
the cross-reference in SeriesModel.__and__ to a lazy import to avoid a
circular import with ParallelModel.
