+++ surpyval/experimental/__init__.py
@@ -1,1 +1,3 @@
 from .forest import RandomSurvivalForest, SurvivalTree
+from .parallel import ParallelModel
+from .series import SeriesModel
